Контроллер Circle

+++ src/Circles.Domain/Circle.cs
@@ -1,3 +1,3 @@
 ﻿namespace Circles.Domain;
 
-public record Circle(int Id, string Description, int x, int y);
+public record Circle(int Id, string Description, long x, long y);
